Add item dont have technical validation  to post new item on a order.Correction name of the field on waiter datatable exibition

+++ app/Http/Services/Stock/StockServiceRepository.php
@@ -177,27 +177,32 @@ class StockServiceRepository implements StockServiceInterFace
     /**
      * @throws \Exception
      */
-    public static function SetItemSaldoZeroException(string $tableNumber): void
+    public static function SetItemSaldoZeroException(string $tableNumber = null, string $menuitem = null): void
     {
-        $cart = Cart::where('tableNumber', $tableNumber)->get();
-        foreach ($cart as $item)
-        {
-            $item_sheets = Technicalfiche::where("itemID", $item->item_id)->get();
-            if (count($item_sheets) < 1){
-                Cart::where('tableNumber', $tableNumber)
-                    ->delete();
-                throw new Exception("Ação não pode ser concluida. O item não tem ficha tecnica");
-            }
-            foreach ($item_sheets as $sheet)
-            {
-                $productID = $sheet->productID ?? 19;
-                $saldo = Saldo::where('productID', $productID)->first();
-                $inventoty = $saldo->saldoFinal ?? 0;
-
-                if (!$inventoty|| $inventoty <= 0){
-                    Cart::where('tableNumber', $tableNumber)->delete();
-                    throw new Exception("Ação não pode ser concluida. O pedido contem item com estoque zerado. {$productID}");
+        if (is_null($menuitem) && !is_null($tableNumber)) {
+            $cart = Cart::where('tableNumber', $tableNumber)->get();
+            foreach ($cart as $item) {
+                $item_sheets = Technicalfiche::where("itemID", $item->item_id)->get();
+                if (count($item_sheets) < 1) {
+                    Cart::where('tableNumber', $tableNumber)
+                        ->delete();
+                    throw new Exception("Ação não pode ser concluída. O item não tem ficha tecnica");
                 }
+                foreach ($item_sheets as $sheet) {
+                    $productID = $sheet->productID ?? 19;
+                    $saldo = Saldo::where('productID', $productID)->first();
+                    $inventoty = $saldo->saldoFinal ?? 0;
+
+                    if (!$inventoty || $inventoty <= 0) {
+                        Cart::where('tableNumber', $tableNumber)->delete();
+                        throw new Exception("Ação não pode ser concluida. O pedido contem item com estoque zerado. {$productID}");
+                    }
+                }
+            }
+        }else {
+            $item_sheets = Technicalfiche::where("itemID", $menuitem)->exists();
+            if (!$item_sheets) {
+                throw new Exception("Ação não pode ser concluída. O item não tem ficha tecnica");
             }
         }
     }
